🚨 test(refactor): use compiled version by pika to import and test repo
