+++ src/index.js
@@ -1,3 +1,5 @@
+module.exports = createWebhooksApi;
+
 const createEventHandler = require("./event-handler");
 const middleware = require("./middleware/middleware");
 const sign = require("./sign");
@@ -26,10 +28,3 @@ function createWebhooksApi(options) {
   };
 }
 
-module.exports = {
-  createWebhooksApi,
-  middleware,
-  sign,
-  verify,
-  verifyAndReceive,
-};
